pyproject.toml
[tool.poetry]
name = "[insert-project-name-here]"
version = "0.1.0"
description = ""
authors = ["farooq-azam-khan <khanazamfarooq99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-dotenv = "^1.0.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^1.3.0" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
